# Add Kerl state and let it implement hash.Hash