--- a/signing/legacy/signing.go
+++ b/signing/legacy/signing.go
@@ -12,6 +12,9 @@ import (
 	. "github.com/iotaledger/iota.go/trinary"
 )
 
+// the default SpongeFunction creator
+var defaultCreator = func() SpongeFunction { return kerl.NewKerl() }
+
 // Subseed takes a seed and an index and returns the given subseed.
 // Optionally takes the SpongeFunction to use. Default is Kerl.
 func Subseed(seed Trytes, index uint64, spongeFunc ...SpongeFunction) (Trits, error) {
@@ -28,7 +31,7 @@ func Subseed(seed Trytes, index uint64, spongeFunc ...SpongeFunction) (Trits, er
 
 	incrementedSeed := AddTrits(trits, IntToTrits(int64(index)))
 
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	err = h.Absorb(incrementedSeed)
@@ -46,7 +49,7 @@ func Subseed(seed Trytes, index uint64, spongeFunc ...SpongeFunction) (Trits, er
 // Key computes a new private key from the given subseed using the given security level.
 // Optionally takes the SpongeFunction to use. Default is Kerl.
 func Key(subseed Trits, securityLevel SecurityLevel, spongeFunc ...SpongeFunction) (Trits, error) {
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	keyLength := KeyFragmentLength * int(securityLevel)
@@ -89,7 +92,7 @@ func Digests(key Trits, spongeFunc ...SpongeFunction) (Trits, error) {
 	digests := make(Trits, HashTrinarySize)
 	buf := make(Trits, HashTrinarySize)
 
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	// iterate through each key fragment
@@ -136,7 +139,7 @@ func Digests(key Trits, spongeFunc ...SpongeFunction) (Trits, error) {
 // Address generates the address trits from the given digests.
 // Optionally takes the SpongeFunction to use. Default is Kerl.
 func Address(digests Trits, spongeFunc ...SpongeFunction) (Trits, error) {
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	if err := h.Absorb(digests); err != nil {
@@ -173,7 +176,7 @@ func SignatureFragment(bundleHash Trits, keyFragment Trits, startOffset int, spo
 	keyLength := int(secLvl) * int(ISSKeyLength)
 	sigFrag := make(Trits, keyLength)
 
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	sig := make(Trits, HashTrinarySize)
@@ -220,7 +223,7 @@ func Digest(bundleHash []int8, signatureFragment Trits, startOffset int, spongeF
 	sigFrag := make(Trits, sigLength)
 	copy(sigFrag, signatureFragment[:sigLength])
 
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	dig := make(Trits, HashTrinarySize)