--- a/signing/signing.go
+++ b/signing/signing.go
@@ -11,6 +11,9 @@ import (
 	. "github.com/iotaledger/iota.go/trinary"
 )
 
+// the default SpongeFunction creator
+var defaultCreator = func() SpongeFunction { return kerl.NewKerl() }
+
 // Subseed takes a seed and an index and returns the given subseed.
 // Optionally takes the SpongeFunction to use. Default is Kerl.
 func Subseed(seed Trytes, index uint64, spongeFunc ...SpongeFunction) (Trits, error) {
@@ -27,7 +30,7 @@ func Subseed(seed Trytes, index uint64, spongeFunc ...SpongeFunction) (Trits, er
 
 	incrementedSeed := AddTrits(trits, IntToTrits(int64(index)))
 
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	err = h.Absorb(incrementedSeed)
@@ -45,7 +48,7 @@ func Subseed(seed Trytes, index uint64, spongeFunc ...SpongeFunction) (Trits, er
 // Key computes a new private key from the given subseed using the given security level.
 // Optionally takes the SpongeFunction to use. Default is Kerl.
 func Key(subseed Trits, securityLevel SecurityLevel, spongeFunc ...SpongeFunction) (Trits, error) {
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	if err := h.Absorb(subseed); err != nil {
@@ -75,7 +78,7 @@ func Digests(key Trits, spongeFunc ...SpongeFunction) (Trits, error) {
 	digests := make(Trits, fragments*HashTrinarySize)
 	buf := make(Trits, HashTrinarySize)
 
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	// iterate through each key fragment
@@ -120,7 +123,7 @@ func Digests(key Trits, spongeFunc ...SpongeFunction) (Trits, error) {
 // Address generates the address trits from the given digests.
 // Optionally takes the SpongeFunction to use. Default is Kerl.
 func Address(digests Trits, spongeFunc ...SpongeFunction) (Trits, error) {
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	if err := h.Absorb(digests); err != nil {
@@ -173,7 +176,7 @@ func SignatureFragment(normalizedBundleHashFragment Trits, keyFragment Trits, sp
 	sigFrag := make(Trits, len(keyFragment))
 	copy(sigFrag, keyFragment)
 
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	for i := 0; i < KeySegmentsPerFragment; i++ {
@@ -201,7 +204,7 @@ func SignatureFragment(normalizedBundleHashFragment Trits, keyFragment Trits, sp
 // Digest computes the digest derived from the signature fragment and normalized bundle hash.
 // Optionally takes the SpongeFunction to use. Default is Kerl.
 func Digest(normalizedBundleHashFragment []int8, signatureFragment Trits, spongeFunc ...SpongeFunction) (Trits, error) {
-	h := GetSpongeFunc(spongeFunc, kerl.NewKerl)
+	h := GetSpongeFunc(spongeFunc, defaultCreator)
 	defer h.Reset()
 
 	buf := make(Trits, HashTrinarySize)