# Description of change

- `Kerl` now implements `hash.Hash`. This allows much more efficient hashing of binary data if no ternary representation is required. (This is the case most of the time.)
- Introduce an internal state for `Kerl`:
  - A buffer to store the hash sum without additional memory allocation.
  - Multiple calls of squeeze require a re-initialization of Keccak. This is now only performed on the beginning of a second squeeze instead of the end of every squeeze.  
- Unroll bit flipping and compute it on `uint64` instead of `byte`.
- Added ginkgo tests for all new functions.

On the tested machine this leads to a speedup for all Kerl related functions. This has been tested and verified with a few artificial benchmarks:

### `amd64`
```
name                         old time/op  new time/op  delta
KerlTrytes-48                2.29µs ± 7%  2.08µs ± 6%   -9.21%
GenerateAddress-48           6.17ms ± 8%  5.64ms ± 6%   -8.59%
ValidateBundleSignatures-48  1.12ms ± 5%  1.00ms ± 4%  -10.86%
```

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

Add new test-cases and verified result on existing tests.

## Change checklist

- [x] My code follows the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests using ginkgo that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes